Remove long extensions

+++ shared/src/androidMain/kotlin/fr/outadoc/justchatting/feature/chat/presentation/mobile/StreamInfo.kt
@@ -63,7 +63,7 @@ internal fun StreamInfo(
                 text = pluralStringResource(
                     R.plurals.viewers,
                     stream.viewerCount.toInt(),
-                    stream.viewerCount.formatNumber(),
+                    stream.viewerCount.toInt().formatNumber(),
                 ),
             )
         }
